refactor(k8s-client)!: replace facade factory with static KubernetesApiClientFacade::create()
BREAKING CHANGE: KubernetesApiClientFacadeFactory is removed; build the facade via the static named
constructor KubernetesApiClientFacade::create(KubernetesApiClient, LoggerInterface, array $extraClients).
Also drops the misleading @internal from KubernetesApiClient (it is the public construction seam) and
documents $extraClients + a Symfony wiring recipe in the README.
