+++ libs/k8s-client/tests/KubernetesApiClientFacadeCreateTest.php
@@ -2,47 +2,41 @@
 
 declare(strict_types=1);
 
-namespace Keboola\K8sClient\Tests\ClientFactory;
+namespace Keboola\K8sClient\Tests;
 
 use Keboola\K8sClient\ApiClient\ApiClientInterface;
-use Keboola\K8sClient\ClientFactory\KubernetesApiClientFacadeFactory;
+use Keboola\K8sClient\ClientFactory\StaticKubernetesApiClientFactory;
 use Keboola\K8sClient\KubernetesApiClient;
 use Keboola\K8sClient\KubernetesApiClientFacade;
-use Keboola\K8sClient\Tests\FakeCrdModel;
 use Kubernetes\Model\Io\K8s\Apimachinery\Pkg\Apis\Meta\V1\Patch;
 use Monolog\Logger;
 use PHPUnit\Framework\TestCase;
 use Retry\RetryProxy;
 
-class KubernetesApiClientFacadeFactoryTest extends TestCase
+class KubernetesApiClientFacadeCreateTest extends TestCase
 {
     public function testCreateBuildsFacadeFromApiClient(): void
     {
-        $apiClient = new KubernetesApiClient(new RetryProxy(), 'my-namespace');
-        $factory = new KubernetesApiClientFacadeFactory(new Logger('test'));
-
-        $facade = $factory->create($apiClient);
+        $facade = KubernetesApiClientFacade::create($this->apiClient(), new Logger('test'));
 
         self::assertInstanceOf(KubernetesApiClientFacade::class, $facade);
     }
 
     public function testCreateThreadsExtraClientsThroughToFacade(): void
     {
-        $apiClient = new KubernetesApiClient(new RetryProxy(), 'my-namespace');
-        $factory = new KubernetesApiClientFacadeFactory(new Logger('test'));
-
         $extraClient = $this->createMock(ApiClientInterface::class);
 
-        $facade = $factory->create($apiClient, [FakeCrdModel::class => $extraClient]);
+        $facade = KubernetesApiClientFacade::create(
+            $this->apiClient(),
+            new Logger('test'),
+            [FakeCrdModel::class => $extraClient],
+        );
 
         self::assertSame($extraClient, $facade->client(FakeCrdModel::class));
     }
 
     public function testCreateThreadsExtraClientMergePatchRouting(): void
     {
-        $apiClient = new KubernetesApiClient(new RetryProxy(), 'my-namespace');
-        $factory = new KubernetesApiClientFacadeFactory(new Logger('test'));
-
         $model = new FakeCrdModel(['metadata' => ['name' => 'thing-1'], 'spec' => ['size' => 2]]);
 
         $extraClient = $this->createMock(ApiClientInterface::class);
@@ -56,8 +50,27 @@ class KubernetesApiClientFacadeFactoryTest extends TestCase
                 return $model;
             });
 
-        $facade = $factory->create($apiClient, [FakeCrdModel::class => $extraClient]);
+        $facade = KubernetesApiClientFacade::create(
+            $this->apiClient(),
+            new Logger('test'),
+            [FakeCrdModel::class => $extraClient],
+        );
 
         self::assertSame($model, $facade->mergePatch($model));
+    }
+
+    /**
+     * Build a client through a real factory so the global KubernetesRuntime\Client is configured —
+     * constructing the facade's core API clients requires it.
+     */
+    private function apiClient(): KubernetesApiClient
+    {
+        return (new StaticKubernetesApiClientFactory(
+            new RetryProxy(),
+            'https://example.test',
+            'token',
+            __DIR__ . '/fixtures/ca.crt',
+            'my-namespace',
+        ))->createApiClient();
     }
 }
